Suite#info should not append a spec
Given tools that offer a more sophisticated compile-time step to prepare
the suites, e.g. canter, we should allow there to be a generic tagging
mechanism so we can add infomation at compile time that can be accessed
at runtime, e.g. source code location information.

We'll appropriate Suite#info for this because the previous behavior of
appending a suite was unwanted anyway.