Factories (#6)
* suite and spec methods are now factories

This is the first part of a larger change. In itself it is a large
change. Now, methods like .it() and .describe() are factories which
return the suite or spec they are creating, not the instance itself.

It's more flexible, but comes at the expense of added boilerplate. Now,
you have to traverse back up to the parent if you want to add a sibling
node, but you can build off the child now, which is more expressive.

The second part of this change will be to reify hooks and change the API
so hooks work the same way

* hooks are now factories

this is the second half of the big change described in the previous
commit. It reifies hooks and iumplements the same idea as specs, but for
hooks, and hooks are now a subclass of specs.

* 1.1.0